Implemented Heap_4
changed the heap 2 implementantion from lastr commit to the heap_4
Problems:
-understanding the code

+++ rdb/Kernel/include/memInter.h
@@ -1,9 +1,9 @@
 #include <stdint.h>
 #include <stddef.h>
 
-void malloc(int wanted_size, void ** ret_val);
+void malloc(size_t wanted_size, void ** ret_val);
 
 void free( void * p);
 
-void getFreeHeapSize(int * resp);
+void getFreeHeapSize(size_t * resp);
