+++ rdb/Kernel/system.c
@@ -1,5 +1,6 @@
 #include <system.h>
 #include <stdint.h>
+#include <stddef.h>
 #include <memInter.h>
 
 static void getMemory(uint64_t * pos, uint64_t * mem_buffer, unsigned int dim) {
@@ -14,13 +15,13 @@ int sys_system(void * option, void * arg1, void * arg2, void * arg3) {
 		getMemory((uint64_t *) arg1, (uint64_t *) arg2, (uint64_t) arg3);
 		break;
 	case 1:
-		malloc( (int) arg1, (void **) arg2);
+		malloc( (size_t) arg1, (void **) arg2);
 		break; 
 	case 2:	
 		free( (void *) arg1);
 		break;	
 	case 3:
-		getFreeHeapSize( (int*) arg1 );
+		getFreeHeapSize( (size_t*) arg1 );
 		break;	
 	return 0;
 	}
